+++ packages/core/src/typography/types.ts
@@ -1,23 +1,15 @@
 import { FC, ReactNode } from 'react'
-import {
-    NumericPositionSpec,
-    SvgElementVariantProps,
-    SvgElementProps,
-    LocationProps,
-} from '../general'
+import { NumericPositionSpec, SvgElementVariantProps, LocationProps } from '../general'
 
-export interface TextContentProps extends Pick<SvgElementProps, 'style' | 'className'> {
-    /** content */
-    children?: ReactNode
-}
-
-export interface TypographyProps extends SvgElementVariantProps, TextContentProps {
+export interface TypographyProps extends SvgElementVariantProps {
     /** position (absolute coordinates) */
     position?: NumericPositionSpec
     /** variant */
     variant?: 'default' | 'title' | 'subtitle' | 'axisLabel' | 'tickLabel' | string
     /** rotation (degrees) */
     angle?: number
+    /** content */
+    children?: ReactNode
 }
 
 export type LabelProps = TypographyProps & LocationProps
@@ -26,7 +18,7 @@ export interface CounterProps extends LabelProps {
     /** number of decimal places */
     nDecimalPlaces?: number
     /** format */
-    format?: (v: number) => string | ReactNode
+    format?: (v: number) => ReactNode
     /** component to replace default text */
-    component?: FC<TextContentProps>
+    component?: FC<LabelProps>
 }